Update youtube.js : Bot can play music from YouTube Playlists
You can now paste a YouTube playlist link and the bot will add all the songs from that playlist into the queue (songs that already exist in the queue will stay and playlist songs will be appended at the end of the queue). Then, the bot begin to play if not already doing so.