Added translate.js : Fun Translation feature
This is a new feature for the bot which takes in user input text and translates this text depending on which translator they choose. These are just fictional/fun languages that a user can choose from. This feature is functional but unfortunately there is a rate limit of 5 requests per hour without a paid subscription.